fix: Resolve clang compiler warnings

+++ core/io/resource_importer.cc
@@ -61,7 +61,7 @@ namespace ho {
             return nullptr;
         }
         TopologicalSort(imported->scene->mRootNode, imported->flatted_scene);
-        for (uint32_t i = 0; i < imported->flatted_scene.size(); i++) {
+        for (size_t i = 0; i < imported->flatted_scene.size(); i++) {
             imported->node_to_index[imported->flatted_scene[i]] = i;
         }
         return imported;
@@ -99,7 +99,7 @@ namespace ho {
             for (int y = 0; y < height; ++y) {
                 for (int x = 0; x < width; ++x) {
                     const aiTexel& texel = aitex->pcData[y * width + x];
-                    int idx = (y * width + x) * num_channels;
+                    size_t idx = static_cast<size_t>((y * width + x) * num_channels);
                     decoded[idx + 0] = texel.r;
                     decoded[idx + 1] = texel.g;
                     decoded[idx + 2] = texel.b;
@@ -142,7 +142,7 @@ namespace ho {
         if (root == nullptr) {
             return;
         }
-        for (unsigned int i = 0; i < root->mNumChildren; i++) {
+        for (size_t i = 0; i < static_cast<size_t>(root->mNumChildren); i++) {
             TopologicalSort(root->mChildren[i], flatted_scene);
         }
         flatted_scene.push_front(root);
